requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
attrs=19.3.0=py_0
blas=1.0=mkl
ca-certificates=2020.1.1=0
certifi=2020.4.5.1=py36_0
cloudpickle=1.4.1=py_0
cudatoolkit=10.2.89=hfd86e86_1
cxxfilt=0.2.1=py36h831f99a_1
cycler=0.10.0=py36_0
cytoolz=0.10.1=py36h7b6447c_0
dask-core=2.17.2=py_0
dbus=1.13.14=hb2f20db_0
decorator=4.4.2=py_0
efficientnet-pytorch=0.6.3=pypi_0
expat=2.2.6=he6710b0_0
fontconfig=2.13.0=h9420a91_0
freetype=2.9.1=h8a8886c_1
future=0.18.2=pypi_0
git=2.23.0=pl526hacde149_0
glib=2.63.1=h3eb4bd4_1
gst-plugins-base=1.14.0=hbbd80ab_1
gstreamer=1.14.0=hb31296c_0
icu=58.2=he6710b0_3
imageio=2.8.0=py_0
importlib-metadata=1.6.1=py36h9f0ad1d_0
importlib_metadata=1.6.1=0
intel-openmp=2020.1=217
joblib=0.15.1=py_0
jpeg=9b=h024ee3a_2
kiwisolver=1.2.0=py36hfd86e86_0
krb5=1.17.1=h173b8e3_0
ld_impl_linux-64=2.33.1=h53a641e_7
libcurl=7.69.1=h20c2e04_0
libedit=3.1.20181209=hc058e9b_0
libffi=3.3=he6710b0_1
libgcc-ng=9.1.0=hdf63c60_0
libgfortran-ng=7.3.0=hdf63c60_0
libpng=1.6.37=hbc83047_0
libssh2=1.9.0=h1ba5d50_1
libstdcxx-ng=9.1.0=hdf63c60_0
libtiff=4.1.0=h2733197_1
libuuid=1.0.3=h1bed415_2
libxcb=1.13=h1bed415_1
libxml2=2.9.10=he19cac6_1
lz4-c=1.9.2=he6710b0_0
matplotlib=3.1.3=py36_0
matplotlib-base=3.1.3=py36hef1b27d_0
mkl=2020.1=217
mkl-service=2.3.0=py36he904b0f_0
mkl_fft=1.0.15=py36ha843d7b_0
mkl_random=1.1.1=py36h0573a6f_0
more-itertools=8.3.0=py_1
ncurses=6.2=he6710b0_1
networkx=2.4=py_0
ninja=1.9.0=py36hfd86e86_0
numpy=1.18.1=py36h4f9e942_0
numpy-base=1.18.1=py36hde5b4d6_1
nvidia-apex=0.1=py36h14a324e_1
olefile=0.46=py36_0
openssl=1.1.1g=h7b6447c_0
packaging=20.4=pyh9f0ad1d_0
pandas=1.0.3=py36h0573a6f_0
pcre=8.43=he6710b0_0
perl=5.26.2=h14c3975_0
pillow=7.1.2=py36hb39fc2d_0
pip=20.0.2=py36_3
pluggy=0.13.1=py36h9f0ad1d_1
py=1.8.1=py_0
pyparsing=2.4.7=py_0
pyqt=5.9.2=py36h05f1152_2
pytest=5.4.3=py36h9f0ad1d_0
python=3.6.10=h7579374_2
python-dateutil=2.8.1=py_0
python_abi=3.6=1_cp36m
pytorch=1.5.0=py3.6_cuda10.2.89_cudnn7.6.5_0
pytz=2020.1=py_0
pywavelets=1.1.1=py36h7b6447c_0
pyyaml=5.3.1=py36h8c4c3a4_0
qt=5.9.7=h5867ecd_1
readline=8.0=h7b6447c_0
scikit-image=0.16.2=py36h0573a6f_0
scikit-learn=0.22.1=py36hd81dba3_0
scipy=1.4.1=py36h0b6359f_0
setuptools=47.1.1=py36_0
sip=4.19.8=py36hf484d3e_0
six=1.15.0=py_0
sqlite=3.31.1=h62c20be_1
tk=8.6.8=hbc83047_0
toolz=0.10.0=py_0
torchvision=0.6.0=py36_cu102
tornado=6.0.4=py36h7b6447c_1
tqdm=4.46.0=py_0
warmup-scheduler=0.3.2=pypi_0
wcwidth=0.2.4=pyh9f0ad1d_0
wheel=0.34.2=py36_0
xz=5.2.5=h7b6447c_0
yaml=0.2.5=h516909a_0
zipp=3.1.0=py_0
zlib=1.2.11=h7b6447c_3
zstd=1.4.4=h0b5b093_3
